+++ lib/index.js
@@ -1,27 +1,12 @@
 var recast = require('recast');
 var types = recast.types;
-var through = require('through');
 var b = recast.types.builders;
 var n = recast.types.namedTypes;
-
-function Visitor() {
-  types.PathVisitor.call(this);
-}
-Visitor.prototype = Object.create(types.PathVisitor.prototype);
-Visitor.prototype.constructor = Visitor;
-
-Visitor.prototype.visitProperty = function(prop) {
-  if (n.ObjectExpression.check(prop.parent.node)) {
-    if (prop.node.shorthand) {
-      prop.node.shorthand = false;
-    }
-  }
-
-  this.traverse(prop);
-};
+var through = require('through');
+var Visitor = require('./visitor');
 
 function transform(ast) {
-  return types.visit(ast, new Visitor());
+  return types.visit(ast, Visitor.visit);
 }
 
 function compile(code, options) {
